Rename captureDdgGlobals
