+++ src/captureDdgGlobals.js
